Update OperationIds for human readable labels (#185)
Ticket:
https://maxioevolution.atlassian.net/browse/DE-312

Type of change:
 internal

What
This PR updates the OAS `operationId`s with human readable labels

How
operationId is optional property for an operation but it’s used for method names in generated SDK.

e.g. post-portal-customers-customer_id-enable.json → postPortalCustomersCustomerIdEnableJson - this name is complicated, is postfixed with Json and is meaningless. Better name would be e.g. enableBillingPortal.

So these operationIds need to be meaningful as they will be most visible and most used part in SDK.